Move Generic request types to avoid circular import
